Create Angry Professor Hacker Rank.py
A Discrete Mathematics professor has a class of students. Frustrated with their lack of discipline, he decides to cancel class if fewer than some number of students are present when class starts. Arrival times go from on time (arrivalTime=<0) to arrived late (arrivalTime>0).
Given the arrival time of each student and a threshhold number of attendees, determine if the class is canceled.

Input Format:
The first line of input contains t , the number of test cases.
Each test case consists of two lines.
The first line has two space-separated integers, N and K , the number of students (size of  a) and the cancellation threshold. 
The second line contains  N space-separated integers (a1,a2,a3,......aN) describing the arrival times for each student.